Prepare multi-tenancy: fresh schema from migrations, configurable branding and mail
- Recreate the legacy tables (tb_members, tb_user, ...) in a guarded baseline
  migration so an empty database can be migrated from scratch; make the
  MySQL-only duty plan migrations driver-agnostic and skip legacy data imports
  when the legacy source tables are missing.
- Stop seeding club-specific settings in migrations.
- Move logo, letterhead, signatories and letter place into the settings
  (new "Briefpapier & E-Mail" page, branding disk, public logo route); import
  the existing Angerberg values and files once on existing installations.
- Apply the letterhead in memory instead of via temp files; skip it when none
  is configured.
- Optional club sender, IMAP account (encrypted password) and birthday
  recipient group in the settings, falling back to .env; IMAP no longer reads
  env() directly.
- Fix UserFactory for tb_user, repair the example feature test, add tests.

+++ database/migrations/2026_09_15_213215_add_external_contact_to_dutyplan_assignments_table.php
@@ -15,7 +15,11 @@ return new class extends Migration
 
             $table->index('externalContactID');
 
-            $table->unsignedBigInteger('memberID')
+            /*
+             * Typ bleibt signed int wie tb_members.memberID — sonst scheitert
+             * MySQL am bestehenden Fremdschlüssel fk_dutyplan_assignment_member.
+             */
+            $table->integer('memberID')
                 ->nullable()
                 ->change();
         });
@@ -27,7 +31,7 @@ return new class extends Migration
             $table->dropIndex(['externalContactID']);
             $table->dropColumn('externalContactID');
 
-            $table->unsignedBigInteger('memberID')
+            $table->integer('memberID')
                 ->nullable(false)
                 ->change();
         });
